pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "magiceye"
version = "0.1.0"
description = "A command-line tool to generate autostereograms (magic eye images)."
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
  "mediapy",
  "absl-py",
  "chex>=0.1.89",
  "einops>=0.8.1",
  "numpy",
  "depth-pro @ git+https://github.com/apple/ml-depth-pro.git",
  "matplotlib>=3.10.3",
  "scipy>=1.16.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.4.1",
]

[project.scripts]
magiceye = "magiceye.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
